+++ onlyoffice-nuxeo-core/src/main/java/org/onlyoffice/utils/UrlManager.java
@@ -31,4 +31,5 @@ public interface UrlManager {
     public String getContentUrl(WebContext ctx, DocumentModel model);
     public String getContentUrl(OperationContext ctx, DocumentModel model);
     public String getCallbackUrl(WebContext ctx, DocumentModel model);
+    public String getGobackUrl(WebContext ctx, DocumentModel model);
 }