+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ## Added
 - ability to create documents
 
+## Changed
+- document server v6.0 and earlier is no longer supported
+
 ## 2.0.0
 ## Added
 - support for version 2021